# frontend/openapi/job-hunt.openapi.json
{
  "components": {
    "schemas": {
...
      "ActionItemKind": {
        "enum": [
          "review_and_prepare_application"
        ],
        "title": "ActionItemKind",
        "type": "string"
      },
      "ActionItemResponse": {
        "additionalProperties": false,
        "properties": {
          "application_id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{0,31}$",
            "title": "Application Id",
            "type": "string"
          },
          "cancelled_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancelled At"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "due_on": {
            "format": "date",
            "title": "Due On",
            "type": "string"
          },
          "id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{0,31}$",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/ActionItemKind"
          },
          "status": {
            "$ref": "#/components/schemas/ActionItemStatus"
          },
          "title": {
            "maxLength": 240,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "minimum": 1.0,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "version",
          "application_id",
          "kind",
          "status",
          "title",
          "due_on",
          "created_at",
          "updated_at"
        ],
        "title": "ActionItemResponse",
        "type": "object"
      },
      "ActionItemStatus": {
        "enum": [
          "open",
          "completed",
          "cancelled"
        ],
        "title": "ActionItemStatus",
        "type": "string"
      },
      "ApplicationActivityEventResponse": {
        "additionalProperties": false,
        "properties": {
          "action_item_id": {
            "anyOf": [
              {
                "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{0,31}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Action Item Id"
          },
          "application_id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{0,31}$",
            "title": "Application Id",
            "type": "string"
          },
          "event_type": {
            "$ref": "#/components/schemas/ApplicationActivityEventType"
          },
          "from_stage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApplicationStage"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{0,31}$",
            "title": "Id",
            "type": "string"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "sequence_number": {
            "minimum": 1.0,
            "title": "Sequence Number",
            "type": "integer"
          },
          "to_stage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApplicationStage"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "application_id",
          "sequence_number",
          "event_type",
          "occurred_at"
        ],
        "title": "ApplicationActivityEventResponse",
        "type": "object"
      },
      "ApplicationActivityEventType": {
        "enum": [
          "application_created"
        ],
        "title": "ApplicationActivityEventType",
        "type": "string"
      },
      "ApplicationActivityListResponse": {
        "additionalProperties": false,
        "properties": {
          "data_source": {
            "const": "database",
            "default": "database",
            "title": "Data Source",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ApplicationActivityEventResponse"
            },
            "maxItems": 500,
            "title": "Items",
            "type": "array"
          }
        },
        "title": "ApplicationActivityListResponse",
        "type": "object"
      },
      "ApplicationDetailResponse": {
        "additionalProperties": false,
        "properties": {
          "activity": {
            "items": {
              "$ref": "#/components/schemas/ApplicationActivityEventResponse"
            },
            "maxItems": 500,
            "minItems": 1,
            "title": "Activity",
            "type": "array"
          },
          "application": {
            "$ref": "#/components/schemas/ApplicationSummary"
          },
          "data_source": {
            "const": "database",
            "default": "database",
            "title": "Data Source",
            "type": "string"
          }
        },
        "required": [
          "application",
          "activity"
        ],
        "title": "ApplicationDetailResponse",
        "type": "object"
      },
      "ApplicationListResponse": {
        "additionalProperties": false,
        "properties": {
          "data_source": {
            "const": "database",
            "default": "database",
            "title": "Data Source",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ApplicationSummary"
            },
            "maxItems": 50,
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "maxLength": 512,
                "minLength": 1,
                "pattern": "^[A-Za-z0-9_-]+$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total"
        ],
        "title": "ApplicationListResponse",
        "type": "object"
      },
      "ApplicationPostingState": {
        "enum": [
          "open",
          "closed",
          "unknown"
        ],
        "title": "ApplicationPostingState",
        "type": "string"
      },
      "ApplicationPostingSummary": {
        "additionalProperties": false,
        "properties": {
          "canonical_url": {
            "maxLength": 2048,
            "minLength": 1,
            "title": "Canonical Url",
            "type": "string"
          },
          "company": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Company",
            "type": "string"
          },
          "first_party": {
            "title": "First Party",
            "type": "boolean"
          },
          "id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{0,31}$",
            "title": "Id",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/ApplicationPostingState"
          },
          "title": {
            "maxLength": 300,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "company",
          "title",
          "canonical_url",
          "first_party",
          "state"
        ],
        "title": "ApplicationPostingSummary",
        "type": "object"
      },
      "ApplicationStage": {
        "enum": [
          "pursuing"
        ],
        "title": "ApplicationStage",
        "type": "string"
      },
      "ApplicationSummary": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "current_action": {
            "$ref": "#/components/schemas/ActionItemResponse"
          },
          "id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{0,31}$",
            "title": "Id",
            "type": "string"
          },
          "opportunity_id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{0,31}$",
            "title": "Opportunity Id",
            "type": "string"
          },
          "posting": {
            "$ref": "#/components/schemas/ApplicationPostingSummary"
          },
          "pursued_posting_version_id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{0,31}$",
            "title": "Pursued Posting Version Id",
            "type": "string"
          },
          "stage": {
            "$ref": "#/components/schemas/ApplicationStage"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "minimum": 1.0,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "version",
          "opportunity_id",
          "pursued_posting_version_id",
          "stage",
          "posting",
          "current_action",
          "created_at",
          "updated_at"
        ],
        "title": "ApplicationSummary",
        "type": "object"
      },
...
      "ProblemFieldError": {
        "additionalProperties": false,
        "properties": {
          "field": {
            "title": "Field",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "field",
          "message"
        ],
        "title": "ProblemFieldError",
        "type": "object"
      },
      "ProblemResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "field_errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ProblemFieldError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field Errors"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "retryable": {
            "default": false,
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ProblemResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OwnerSessionCookie": {
        "description": "Opaque HttpOnly session issued by POST /api/session.",
        "in": "cookie",
        "name": "job_hunt_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Job Hunt Signal API",
    "version": "0.4.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/applications": {
      "get": {
        "operationId": "list_owner_applications_api_applications_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 512,
                  "minLength": 1,
                  "pattern": "^[A-Za-z0-9_-]+$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Precondition Required"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "OwnerSessionCookie": []
          }
        ],
        "summary": "List Owner Applications",
        "tags": [
          "applications"
        ]
      }
    },
    "/api/applications/{application_id}": {
      "get": {
        "operationId": "get_owner_application_api_applications__application_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{0,31}$",
              "title": "Application Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Precondition Required"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "OwnerSessionCookie": []
          }
        ],
        "summary": "Get Owner Application",
        "tags": [
          "applications"
        ]
      }
    },
    "/api/applications/{application_id}/activity": {
      "get": {
        "operationId": "list_owner_application_activity_api_applications__application_id__activity_get",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{0,31}$",
              "title": "Application Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationActivityListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Precondition Required"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "OwnerSessionCookie": []
          }
        ],
        "summary": "List Owner Application Activity",
        "tags": [
          "applications"
        ]
      }
    },
...
  }
}